Extract backend Args to schemas/backends
Move BackendArgs, OpenAIHTTPBackendArgs, OpenAIWebSocketBackendArgs,
and VLLMPythonBackendArgs into schemas/backends/ subpackage. Original
backend implementation files now import Args from the centralized
schemas module.

Generated-by: Claude Opus 4.6
