+++ src/guidellm/backends/__init__.py
@@ -9,6 +9,8 @@ the abstract Backend base class, OpenAI-compatible HTTP backend, and response
 handlers for processing streaming and non-streaming API responses.
 """
 
+from guidellm.schemas.backends import OpenAIHTTPBackendArgs, VLLMPythonAsyncBackendArgs
+
 from .backend import Backend, BackendArgs
 from .openai import (
     AudioRequestHandler,
@@ -32,12 +34,14 @@ __all__ = [
     "BackendArgs",
     "ChatCompletionsRequestHandler",
     "OpenAIHTTPBackend",
+    "OpenAIHTTPBackendArgs",
     "OpenAIRequestHandler",
     "OpenAIRequestHandlerFactory",
     "OpenAIWebSocketBackend",
     "OpenAIWebSocketBackendArgs",
     "TextCompletionsRequestHandler",
     "VLLMPythonAsyncBackend",
+    "VLLMPythonAsyncBackendArgs",
     "VLLMPythonBatchBackend",
     "VLLMResponseHandler",
 ]
